requirements.txt
pygame>=2.0.0
pyserial>=3.5
RPi.GPIO>=0.7.0
python-dotenv>=0.19.0
pytest>=6.0.0
pytest-cov>=2.10.0
mock>=4.0.0
numpy>=1.19.0